Make missing processor log just debug
Right now, it's not possible to try to configure scabbard at a root project level in a way that's mindful of dependencies. That is to say - it's too late to apply its plugin if we wait until dependencies are resolved, so we're eagerly adding it with the expectation that it will no-op if no annotation processing is run. This comes with an unfortunate side effect due to this noisy log though, and this PR proposes making it a debug log since the project should just no-op without it.